fix(types): some PrometheusConnectionOptions parameters are optional and export everything from index.ts
