+++ examples/nodejs/app.ts
@@ -1,7 +1,8 @@
-import { PrometheusDriver } from '../../dist/prometheus-query.cjs';
+
+import { PrometheusDriver, Alert, Metric, QueryResult } from '../../';
 
 const prom = new PrometheusDriver({
-    endpoint: "http://demo.robustperception.io:9090/",
+    endpoint: 'http://demo.robustperception.io:9090/',
 });
 
 const query = 'up{instance="demo.robustperception.io:9100",job="node"}';
@@ -9,7 +10,7 @@ const query = 'up{instance="demo.robustperception.io:9100",job="node"}';
 
 // last value
 prom.instantQuery(query, null)
-    .then((res) => {
+    .then((res: QueryResult) => {
         console.log("****************", "[instantQuery] Query:", query, "****************")
         console.log("\n");
 
@@ -25,7 +26,7 @@ prom.instantQuery(query, null)
 
 // during past 24h, 1 point every 6 hours
 prom.rangeQuery(query, new Date().getTime() - 24 * 60 * 60 * 1000, new Date(), 6 * 60 * 60)
-    .then((res) => {
+    .then((res: QueryResult) => {
         console.log("****************", "[rangeQuery] Query:", query, "****************");
         console.log("\n");
 
@@ -41,7 +42,7 @@ prom.rangeQuery(query, new Date().getTime() - 24 * 60 * 60 * 1000, new Date(), 6
 
 // list series matching query
 prom.series(query, new Date().getTime() - 24 * 60 * 60 * 1000, new Date())
-    .then((res) => {
+    .then((res: Metric[]) => {
         console.log("****************", "[series] matching:", query, "****************");
         console.log('[series] Series:');
         console.log(res.join('\n'));
@@ -51,7 +52,7 @@ prom.series(query, new Date().getTime() - 24 * 60 * 60 * 1000, new Date())
 
 // list all active alerts
 prom.alerts()
-    .then((res) => {
+    .then((res: Alert[]) => {
         console.log("****************", "[alerts]", "****************");
         console.log(res)
         console.log("\n");
